[Refactor][1/N] Refactor propose function
